Application Programming Interface (APIs)

Documenting the public user interface

Zenith Angle

Insolation.ZenithAngleCalc.daily_zenith_angle — Method

daily_zenith_angle(date::DateTime,
                   latitude::FT) where {FT <: Real}

returns the daily averaged zenith angle and earth-sun distance at a particular latitude on the given date

Insolation.ZenithAngleCalc.daily_zenith_angle — Method

daily_zenith_angle(days_since_equinox::I,
                   obliquity::FT,
                   perihelion::FT,
                   eccentricity::FT,
                   latitude::FT) where {FT <: Real, I <: Int}

returns the daily averaged zenith angle and earth-sun distance at a particular latitude given the days since vernal equinox (defined as March 21), orbital obliquity, longitude of perihelion, and eccentricity

Insolation.ZenithAngleCalc.instantaneous_zenith_angle — Method

instantaneous_zenith_angle(date::DateTime,
                           timezone::FT,
                           longitude::FT,
                           latitude::FT,
                           obliquity::FT,
                           perihelion::FT,
                           eccentricity::FT) where {FT <: Real}

returns the zenith angle and earth-sun distance at a particular longitude and latitude on the given date given orbital parameters: obliquity, longitude of perihelion, and eccentricity

Insolation.ZenithAngleCalc.instantaneous_zenith_angle — Method

instantaneous_zenith_angle(date::DateTime,
                           timezone::FT,
                           longitude::FT,
                           latitude::FT) where {FT <: Real}

returns the zenith angle and earth-sun distance at a particular longitude and latitude on the given date

equations from "Astronomical Algorithms" by Jean Meeus see documentation for details

Insolation

Insolation.InsolationCalc.daily_insolation — Method

daily_insolation(date::DateTime,
                 latitude::FT) where {FT <: Real}

returns the daily averaged insolation at a particular latitude on the given date

Insolation.InsolationCalc.daily_insolation — Method

daily_insolation(days_since_equinox::I,
                 obliquity::FT,
                 perihelion::FT,
                 eccentricity::FT,
                 latitude::FT) where {FT <: Real, I <: Int}

returns the daily averaged insolation at a particular latitude given the days since vernal equinox (defined as March 21), orbital obliquity, longitude of perihelion, and eccentricity

Insolation.InsolationCalc.insolation — Method

insolation(θ::FT,
                 d::FT) where {FT <: Real}

returns the insolation given the zenith angle and earth-sun distance

Insolation.InsolationCalc.instantaneous_insolation — Method

instantaneous_insolation(date::DateTime,
                         timezone::FT,
                         longitude::FT,
                         latitude::FT) where {FT <: Real}

returns the daily averaged insolation at a particular latitude on the given date