Application Programming Interface (APIs)

Documenting the public user interface

Zenith Angle

Insolation.daily_zenith_angle — Method

daily_zenith_angle(date::DateTime,
                   latitude::FT,
                   param_set::APS) where {FT <: Real}

returns the daily averaged zenith angle and earth-sun distance at a particular latitude given the date and orbital parameters obliquity, longitude of perihelion, and eccentricity param_set is an AbstractParameterSet from CLIMAParameters.jl

Insolation.instantaneous_zenith_angle — Method

instantaneous_zenith_angle(date::DateTime,
                           longitude::FT,
                           latitude::FT,
                           param_set::APS) where {FT <: Real}

returns the zenith angle and earth-sun distance at a particular longitude and latitude on the given date (and time UTC) given orbital parameters: obliquity, longitude of perihelion, and eccentricity param_set is an AbstractParameterSet from CLIMAParameters.jl

Insolation

Insolation.insolation — Method

insolation(θ::FT, d::FT, param_set::APS) where {FT <: Real}

returns the insolation given the zenith angle and earth-sun distance param_set is an AbstractParameterSet from CLIMAParameters.jl

Insolation.solar_flux_and_cos_sza — Method

solar_flux_and_cos_sza(date::DateTime,
                  longitude::FT,
                  latitude::FT,
                  param_set::APS) where {FT <: Real}

returns the top-of-atmosphere (TOA) solar flux, i.e. the total solar irradiance (TSI) weighted by the earth-sun distance and cos(solar zenith angle) for input to RRTMGP.jl param_set is an AbstractParameterSet from CLIMAParameters.jl