Application Programming Interface (APIs)

Documenting the public user interface

Orbital Parameters

Insolation.orbital_params — Method

orbital_params(dt::FT) where {FT <: Real}

This function returns the orbital parameters (ϖ, γ, e) at dt (years) since J2000 epoch. Data are read from file and interpolation function are created in init() method. The functions are stored as global variables that are used inside of Insolation.jl. The parameters vary due to Milankovitch cycles. Data from this paper are in the "src/data/INSOL.LA2004.BTL.csv" file.

Zenith Angle

Insolation.daily_zenith_angle — Method

daily_zenith_angle(date::DateTime,
                   latitude::FT,
                   param_set::APS;
                   eot_correction::Bool=true,
                   milankovitch::Bool=true) where {FT <: Real}

Returns the daily averaged zenith angle and earth-sun distance at a particular latitude given the date and orbital parameters obliquity, longitude of perihelion, and eccentricity param_set is an AbstractParameterSet from CLIMAParameters.jl.

eot_correction is an optional Boolean keyword argument that defaults to true when set to true the equation of time correction is turned on. This switch functionality is implemented for easy comparisons with reanalyses.

milankovitch is an optional Boolean keyword argument that defaults to true when set to true the orbital parameters are calculated for the given DateTime, when set to false the orbital parameters at the J2000 epoch from CLIMAParameters are used.

Insolation.instantaneous_zenith_angle — Method

instantaneous_zenith_angle(date::DateTime,
                           longitude::FT,
                           latitude::FT,
                           param_set::APS;
                           eot_correction::Bool=true,
                           milankovitch::Bool=true) where {FT <: Real}

Returns the zenith angle and earth-sun distance at a particular longitude and latitude on the given date (and time UTC) given orbital parameters: obliquity, longitude of perihelion, and eccentricity param_set is an AbstractParameterSet from CLIMAParameters.jl.

eot_correction is an optional Boolean keyword argument that defaults to true when set to true the equation of time correction is turned on. This switch functionality is implemented for easy comparisons with reanalyses.

milankovitch is an optional Boolean keyword argument that defaults to true when set to true the orbital parameters are calculated for the given DateTime when set to false the orbital parameters at the J2000 epoch from CLIMAParameters are used.

Insolation

Insolation.insolation — Method

insolation(θ::FT, d::FT, param_set::APS) where {FT <: Real}

Returns the insolation given the zenith angle and earth-sun distance param_set is an AbstractParameterSet from CLIMAParameters.jl.

Insolation.solar_flux_and_cos_sza — Method

solar_flux_and_cos_sza(date::DateTime,
                  longitude::FT,
                  latitude::FT,
                  param_set::APS) where {FT <: Real}

Returns the top-of-atmosphere (TOA) solar flux, i.e. the total solar irradiance (TSI) weighted by the earth-sun distance and cos(solar zenith angle) for input to RRTMGP.jl param_set is an AbstractParameterSet from CLIMAParameters.jl.